streamlit/combined/requirements.txt
streamlit==1.29.0
tensorflow-cpu==2.15.0
opencv-python-headless==4.8.1.78
numpy==1.24.3
Pillow==10.1.0
plotly==5.18.0